requirements.txt
# Core dependencies
Pillow>=10.0.0
pillow-heif>=0.13.0
FastAPI>=0.104.0
uvicorn[standard]>=0.24.0
Jinja2>=3.1.0
python-multipart>=0.0.6
PyYAML>=6.0
aiofiles>=23.0.0

# Raspberry Pi specific (install on Pi only)
# RPi.GPIO>=0.7.0
# spidev>=3.6
# smbus2>=0.4.0
# gpiozero>=2.0

# Google Photos (optional)
# google-auth>=2.0.0
# google-auth-oauthlib>=1.0.0
# google-api-python-client>=2.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
